requirements.txt
flask
openai
pinecone
pinecone-client
python-dotenv
langchain
langchain_community
langchain_text_splitters
langchain_openai
langchain_pinecone